Exhume the interner
I actually stopped using it in the lexer, and
It now lives out-of-tree:
https://github.com/CAD97/simple-interner
https://crates.io/crates/simple-interner
https://redd.it/8ah8k7

The modal lexer is now zero-copy, so doesn't need it.
The AST will probably use the interner, but I'm still
up in the air about storing &str vs Interned<str>.

The main reason is that Interned<T> _cannot_ be Serialize.
It's based on the invariant that its &str is owned by a Interner<T>.

But for now, nafi-interner is dead code. BEGONE!